requirements.txt
attrs==19.3.0
bcrypt==3.1.7
cached-property==1.5.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
dj-database-url==0.5.0
Django==2.2.9
django-filter==2.2.0
djangorestframework==3.11.0
docker==4.1.0
docker-compose==1.25.3
dockerpty==0.4.1
docopt==0.6.2
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
jsonschema==3.2.0
numpy==1.18.1
pandas==1.0.0
paramiko==2.7.1
psycopg2-binary==2.8.4
pycparser==2.19
PyNaCl==1.3.0
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
texttable==1.6.2
urllib3==1.25.8
websocket-client==0.57.0
whitenoise==5.0.1
zipp==2.1.0
